router: Move test tlsCerts into router/testutils package
